+++ YourFavECommerce/Models/Promotion.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
 
         public string Code { get; set; }
-        public decimal Discount { get; set; }
+        public long Discount { get; set; }
         public DateTime CreateAt { get; set; } = DateTime.UtcNow;
         public DateTime ValidTo { get; set; } = DateTime.UtcNow.AddDays(7);
         public int MaxOfUsage { get; set; } = 100;
