+++ YourFavECommerce/ViewModels/CreateProductVM.cs
@@ -7,8 +7,8 @@
         public IFormFile MainImg { get; set; } = default!;
         public List<IFormFile>? SubImages { get; set; } 
         public List<string>? Colors { get; set; }
-        public decimal Price { get; set; }
-        public decimal Discount { get; set; }
+        public long Price { get; set; }
+        public long Discount { get; set; }
         public int Quantity { get; set; }
         public int CategoryId { get; set; }
         public int BrandId { get; set; }
